refactor: 전반적인 컬러변수 수정 및 아이콘 추가 1. 404아이콘을 추가했습니다 2. main-subtext-color 컬러 변수를추가했습니다. 3. 디자인 수정을위해 변수들을 수정했습니다.
