+++ src/components/splash/Splash.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import splashLogo from '../../assets/full-logo.svg';
+import splashLogo from '../../assets/full-logo-W-new-01.svg';
 import styled, { keyframes } from 'styled-components';
 
 const ImageFadeOut = keyframes`
@@ -13,10 +13,13 @@ const ImageFadeOut = keyframes`
 `;
 
 const SplashImg = styled.img`
-  height: 200px;
+  height: 225px;
+  margin-top: -252px;
+  margin-left: -0.5px;
 `;
 
 const FlexCenter = styled.div`
+  width: 100%;
   height: 100%;
   display: flex;
   justify-content: conter;
@@ -28,9 +31,9 @@ const AniSplash = styled.div`
   left: 0;
   width: 100%;
   height: 100%;
-  background-color: var(--bg-color);
-  animation: 0.5s ease-in-out ${ImageFadeOut} forwards;
-  animation-delay: 0.5s;
+  background-color: var(--main-color);
+  animation: 0.7s ease-in-out ${ImageFadeOut} forwards;
+  animation-delay: 0.7s;
   z-index: 10;
 `;
