+++ src/pages/LoginHomePage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useEffect, useState } from 'react';
 import styled from 'styled-components';
-import logoImg from '../assets/symbol-logo-W.png';
+import logoImg from '../assets/full-logo-01.svg';
 import LoginCard from '../components/login/LoginCard';
 import Splash from '../components/splash/Splash';
 
@@ -12,13 +12,13 @@ const LoginContainer = styled.main`
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  background-color: var(--main-color);
+  background-color: var(--bg-color);
 `;
 
 const LoginLogoImg = styled.img`
-  width: 144px;
-  height: 144px;
-  margin-bottom: 177px;
+  width: 200px;
+  /* margin-bottom: 120px; */
+  margin: 0 auto 120px;
 `;
 
 export default function LoginHomePage(props) {
